shift things around to be a bit clearer
