+++ data_science/apps/palette_api/app/static/javascript.js
